server/requirements.txt
tensorflow

flask
pillow
numpy
inference-sdk
